requirements.txt
openai
fastapi
uvicorn
python-dotenv
mysql-connector-python
